refactor: patch endpoints for lots

+++ src/TrustPay.Application/Lots/Commands/DeleteLot/DeleteLotCommand.cs
@@ -2,6 +2,7 @@
 
 using MediatR;
 using TrustPay.Application.Common.Interfaces;
+using TrustPay.Application.Common.Interfaces.Auth;
 using TrustPay.Application.Common.Interfaces.EntitiesRepo;
 using TrustPay.Domain.Common;
 
@@ -11,11 +12,12 @@ public class DeleteLotCommandHandler : IRequestHandler<DeleteLotCommand, Result>
 {
     private readonly ILotRepository _lotRepository;
     private readonly IUnitOfWork _unitOfWork;
-
-    public DeleteLotCommandHandler(ILotRepository lotRepository, IUnitOfWork unitOfWork)
+    private readonly ICurrentUserService _currentUserService;
+    public DeleteLotCommandHandler(ILotRepository lotRepository, IUnitOfWork unitOfWork,ICurrentUserService currentUserService)
     {
         _lotRepository = lotRepository;
         _unitOfWork = unitOfWork;
+        _currentUserService = currentUserService;
     }
 
     public async Task<Result> Handle(DeleteLotCommand request, CancellationToken cancellationToken)
@@ -25,8 +27,13 @@ public class DeleteLotCommandHandler : IRequestHandler<DeleteLotCommand, Result>
         {
             return Error.NotFound("Lot.NotFound", "Лот не найден.");
         }
+        var currentUser = _currentUserService.UserId;
+        if (lot.UserId != currentUser && !_currentUserService.IsAdmin)
+        {
+            return Result.Failure("Только создатель лота или админ может удалить лот.");
+        }
 
-         _lotRepository.Delete(lot);
+        _lotRepository.Delete(lot);
         await _unitOfWork.SaveChangesAsync(cancellationToken);
 
         return Result.Success();
